🧬 Creation of an 'utils' file
Why
To store the logic of week creation and control so it's easier to share the data between components

How
- Creating of a 'utils.ts' file
- Exporting each function

Tags
On branch days
Changes to be committed:
 modified:   src/components/Days.vue
 new file:   src/components/utils.ts
